requirements.txt
fastapi
uvicorn==0.23.1
pandas
numpy
matplotlib==3.8.0
tensorflow==2.16.2
scikit-learn==1.3.1
Pillow==10.1.0
beautifulsoup4==4.12.2
jinja2
python-multipart